# openapi/agh.json
{
...
    "info": {
        "title": "AGH API",
        "version": "1.0.0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/api/skills": {
            "get": {
                "operationId": "listSkills",
                "parameters": [
                    {
                        "description": "Workspace id or path",
                        "in": "query",
                        "name": "workspace",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "skills": {
                                            "items": {
                                                "properties": {
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "dir": {
                                                        "type": "string"
                                                    },
                                                    "enabled": {
                                                        "type": "boolean"
                                                    },
                                                    "metadata": {
                                                        "additionalProperties": {},
                                                        "type": "object"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "provenance": {
                                                        "nullable": true,
                                                        "properties": {
                                                            "installed_at": {
                                                                "format": "date-time",
                                                                "type": "string"
                                                            },
                                                            "registry": {
                                                                "type": "string"
                                                            },
                                                            "slug": {
                                                                "type": "string"
                                                            },
                                                            "version": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "required": [
                                                            "installed_at",
                                                            "registry",
                                                            "slug",
                                                            "version"
                                                        ],
                                                        "type": "object"
                                                    },
                                                    "source": {
                                                        "type": "string"
                                                    },
                                                    "version": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "description",
                                                    "dir",
                                                    "enabled",
                                                    "name",
                                                    "source"
                                                ],
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": ["skills"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Invalid workspace filter"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Workspace not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Skills registry is not configured"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "List skills for one workspace",
                "tags": ["skills"],
                "x-agh-transports": ["http", "uds"]
            }
        },
        "/api/skills/{name}": {
            "get": {
                "operationId": "getSkill",
                "parameters": [
                    {
                        "description": "Skill name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Workspace id or path",
                        "in": "query",
                        "name": "workspace",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "skill": {
                                            "properties": {
                                                "description": {
                                                    "type": "string"
                                                },
                                                "dir": {
                                                    "type": "string"
                                                },
                                                "enabled": {
                                                    "type": "boolean"
                                                },
                                                "metadata": {
                                                    "additionalProperties": {},
                                                    "type": "object"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "provenance": {
                                                    "nullable": true,
                                                    "properties": {
                                                        "installed_at": {
                                                            "format": "date-time",
                                                            "type": "string"
                                                        },
                                                        "registry": {
                                                            "type": "string"
                                                        },
                                                        "slug": {
                                                            "type": "string"
                                                        },
                                                        "version": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": [
                                                        "installed_at",
                                                        "registry",
                                                        "slug",
                                                        "version"
                                                    ],
                                                    "type": "object"
                                                },
                                                "source": {
                                                    "type": "string"
                                                },
                                                "version": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "description",
                                                "dir",
                                                "enabled",
                                                "name",
                                                "source"
                                            ],
                                            "type": "object"
                                        }
                                    },
                                    "required": ["skill"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Invalid skill lookup"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Skill or workspace not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Skills registry is not configured"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Get one skill definition",
                "tags": ["skills"],
                "x-agh-transports": ["http", "uds"]
            }
        },
        "/api/skills/{name}/content": {
            "get": {
                "operationId": "getSkillContent",
                "parameters": [
                    {
                        "description": "Skill name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Workspace id or path",
                        "in": "query",
                        "name": "workspace",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "content": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["content"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Invalid skill lookup"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Skill or workspace not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Skills registry is not configured"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Get the raw content for one skill",
                "tags": ["skills"],
                "x-agh-transports": ["http", "uds"]
            }
        },
        "/api/skills/{name}/disable": {
            "post": {
                "operationId": "disableSkill",
                "parameters": [
                    {
                        "description": "Skill name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Workspace id or path",
                        "in": "query",
                        "name": "workspace",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": ["ok"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Invalid skill lookup"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Skill or workspace not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Skills registry is not configured"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Disable one skill",
                "tags": ["skills"],
                "x-agh-transports": ["http", "uds"]
            }
        },
...
    },
...
}